+++ src/util/surfacequery.cpp
@@ -20,7 +20,7 @@ constexpr double kPi = 3.14159265358979323846;
 
 }  // namespace
 
-TinSurfaceQuery::TinSurfaceQuery(const std::vector<float>& vertsXyz, const std::vector<std::uint32_t>& indices)
+TinSurfaceQuery::TinSurfaceQuery(const std::vector<double>& vertsXyz, const std::vector<std::uint32_t>& indices)
     : verts_(&vertsXyz), indices_(&indices), index_(BuildTinSpatialIndex(vertsXyz, indices)) {}
 
 bool TinSurfaceQuery::triangleAt(double x, double y, double* outZ, size_t* outTri) const {
